Fixed support for reporting devices with only ipv6.
ethtool.get_active_devices does not consider as up the devices that
have only an IPv6 address set. Also, it throws an IOError exception
when doing get_ipaddr of a device that is up but does not have an
IPv4 address set. Thus, doing the filtering by using the device flags
and getting the IPv4 address in a similar way as how we get the IPv6
one, not only fixes the issue but increases the consistency of the
code.

Change-Id: I92010740da24bb30c7316c268c704d9067b28aae
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>