agent: Include windows files in tar balls
In order to be able providing source tar balls and src zips containing
all possible build targets we're now including all files necessary.
Previously we did not include all files and it was necessary to
checkout the source git repository to be able building windows
targets for the guest agent.
This commit provides the necessary changes to include all necessary
files for building it from source tar balls.

Change-Id: I793b84ac78127593a11ab9894fdbc68fa45d224d
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>