agent: Adding --with-dist for packaging ease
Since not all platforms can fullfill the requirements of the
configure script, but tar balls need to be creatable anyway,
add a --with-dist to configure for being able to create the
distribution tarballs anywhere.

--with-dist on configure now disables the requirement checks.
This is needed as for example F19 or RHEL6/CentOS won't be able
to fulfill the requirements for the GDM plugin.

Change-Id: If1daa4438d9037d47ed4e7ee0be35cb1c2755779
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>