Revert "Remove contra productive _string_check"
This reverts commit b06a6c4c8e61013b7461c6fc946c1530ae47584e.
The previous commit breaks the guest agent badly because the
string type is required to be unicode, this code previously was
converting it most of the times to unicode, however in some cases
it won't do so and causes the data to be an ascii string or even
replaces valid characters even though it should not.

The attempt for fixing this issue will be addressed by a new patch.

Change-Id: I47176ac60e069e7d557312d7047d9bc4b15ac768
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>