win32: Retrieve the active session through enumeration
Using WTSGetActiveConsoleSessionId() does in case of RDP connections give
a session id which is not corresponding to the users session in some cases.
By enumerating through the sessions and retriving the session with the 'State'
'WTSActive' we are able to reliably retrieve the current active session.
This is required in order to get the logged-in user and being able to retrieve
the user token for performing operations in their name.

Change-Id: I569d4b6b80e102b83b7e6e39c12524f2d588ad37
Bug-Url: https://bugzilla.redhat.com/1068750
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>