Implement default configuration values
When adding new config values to the configuration files, it's possible
that due to the installation the configuration files won't get updated,
when they are modified by an user/administrator on that given system.

By using defaults, we can override this problem and ensure a proper
application start, when config values are missing which are mandatory,
to the functionality.

Change-Id: I5ce9d457e826448aad205ce4abda5dc4b0210804
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1100249