Support for mapping image ids to physical devices
This patch uses libudev through ctypes for systems providing
libudev due to the instable nature of sysfs. While on RHEL5
we're using sysfs we cannot for example use it on Fedora 20
with kernel 3.15 since the serial values seem not to be exposed
by sysfs.

This patch extends the disks-usage message with the 'mapping' field.
The value of the mapping field is an object {'$serial': {'name': '$devname'}}

Change-Id: Ie8961b3703703df573dbf2f0fca2143b463ae76b
Bug-Url: https://bugzilla.redhat.com/1127607
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>