win32: OpenKey third parameter must always be 0
The third parameter is a reserved parameter, where the fourth is
the parameter for the requested access rights. Previously the requested
access was passed as the reserved parameter which is wrong, however it
did not fail any testing.
This patch corrects the misusage.

Change-Id: I9fb3747edec2a7ea7270522ac7104748f7021870
Bug-Url: https://bugzilla.redhat.com/1158470
Bug-Url: https://bugzilla.redhat.com/1105056
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>
(cherry picked from commit 93aa6178f79189b41c28312e9c839e8d8045e52f)