win32: Update version information reporting for Win8+ and simplify
With Windows 8.1 the API for retrieving the OS version on Windows has
changed that it won't report a higher version than 6.2 even though it
might be Windows 8.1 or Windows 10 etc

This code change introduces a new way to retrieve the current version by
using RtlGetVersion which still does behave the old way.
Additionally this patch already introduces the Windows 2016 Server detection
as the version for this is known.

Also this patch now removes obsolete and unsupported versions.

Change-Id: I8047842b080140e072b04c2af46ec0e6b8287c51
Bug-Url: https://bugzilla.redhat.com/1279980
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>