win32: Do not attempt to login when the desktop is unlocked
When the desktop is unlocked we should not try to login because we're
calling the 'Secure Attention Sequence' which will open the fullscreen
window with the options to lock the screen, change password etc. This is
a disturbing side effect that should be avoided.

After searching for the best possible solution I came to the conclusion
that the solution introduced in this patch is good enough, even though
it won't work right after starting the agent. It will start working after
the first login or unlock of the screen. That means also that this
functionality won't be working after a restart of the service, since the
agent won't be aware of the fact that it has been only restarted.
I also can't say by default that the screen is unlocked, because that
would break the SSO on startup and other cases. Therefore I assumed
this as the currently 'good enough solution' as this is also a niche
problem because for this feature to be useful the option to take no
action on console disconnect has to be active in the ovirt-engine. The
default is the lock screen behavior which will intiate the locking of
the screen in case of a disconnect of the console.

Bug-Url: https://bugzilla.redhat.com/1365925
Backport-To: 4.1
Change-Id: I52e92dc00eb34f9f6afc98dc4743a7d793de858b
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>