linux: silence some of the noise caused by querying containers
We query the list of containers periodically every few minutes which
spams the logs with messages from PAM and userhelper. Don't run the
containers-list script if there is no docker running to avoid the spam
completely on machines without docker. Also silence the successfull
auths through PAM on machines with docker.

Change-Id: Ie8f71825359077c7519e1224681161891418f172
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1779160